change queue and stack to use singly linked list to use less memory

+++ derekmaffett/lib/data_structures/stack.rb
@@ -2,7 +2,7 @@ module LeetCode
   # Basic stack for general use
   class Stack
     def initialize(*args)
-      @stack = LeetCode::DoubleList.new
+      @stack = LeetCode::SingleList.new
       args.each { |arg| push arg }
     end
